Tweak to use per-platform server URL for HMRs
Android emulator needs to use 10.0.2.2 instead of localhost

This also reinstates the CSP by having the HMR server listen on 0.0.0.0
(this way, when android tries to reach it at 10.0.2.2 instead of
localhost, it will work).

Deleted constants.js and the other errant server css as the server is
not needed for the template.